+++ packages/create-app/enums/index.js
@@ -1,0 +1,1 @@
+export const BLOCKLET_COMMAND = 'blocklet';