+++ packages/create-app/lib/server.js
@@ -1,21 +1,22 @@
 import { $, which } from 'zx';
 import semver from 'semver';
+import { BLOCKLET_COMMAND } from '../enums/index.js';
 
 $.verbose = false;
 
-async function trimServerOutputVersion(output = '', command) {
+export async function trimServerOutputVersion(output = '', command) {
   // 调用 blocklet 命令时，都会在第一行先打印一个 blocklet [command] [version] 的信息，需要把这个信息 trim 掉
   const version = await getServerVersion();
   if (command) {
-    return output?.replace(`blocklet ${command} v${version}\n`, '');
+    return output?.replace(`${BLOCKLET_COMMAND} ${command} v${version}\n`, '');
   }
-  const reg = new RegExp(`blocklet \\S+ v${version}\\n+`, 'g');
-  return output?.replace(reg, '');
+  const reg = new RegExp(`${BLOCKLET_COMMAND} \\S+ v${version}\\n+`, 'g');
+  return output?.replace(reg, '').trim();
 }
 
 export async function checkServerInstalled() {
   try {
-    await which('blocklet');
+    await which(BLOCKLET_COMMAND);
     return true;
   } catch (e) {
     return false;
@@ -24,7 +25,7 @@ export async function checkServerInstalled() {
 
 export async function getServerVersion() {
   try {
-    const { stdout: output } = await $`blocklet --version`;
+    const { stdout: output } = await $`${BLOCKLET_COMMAND} --version`;
     return output.trim();
   } catch (e) {
     return '0.0.0';
@@ -33,7 +34,7 @@ export async function getServerVersion() {
 
 export async function getServerStatus() {
   try {
-    const { stdout: output } = await $`blocklet server status`;
+    const { stdout: output } = await $`${BLOCKLET_COMMAND} server status`;
     const [matchStr] = output.match(/Blocklet Server status:[\s\S]*?\n/g) || [];
     const status = matchStr.replace(/Blocklet Server status:\s*(\S+)\s*\S*\n*[\s\S]*/gm, '$1');
     return status.toLowerCase();
@@ -54,7 +55,7 @@ export async function checkSatisfiedVersion() {
 
 export async function getServerDirectory() {
   try {
-    const { stdout: output } = await $`blocklet server status`;
+    const { stdout: output } = await $`${BLOCKLET_COMMAND} server status`;
     const [matchStr] = output.match(/Blocklet Server Data Directory:[\s\S]*?\n/gm) || [];
     if (!matchStr) return null;
 
@@ -67,8 +68,8 @@ export async function getServerDirectory() {
 
 export async function getUserInfo() {
   try {
-    const { stdout: name } = await $`blocklet config get name`;
-    const { stdout: email } = await $`blocklet config get email`;
+    const { stdout: name } = await $`${BLOCKLET_COMMAND} config get name`;
+    const { stdout: email } = await $`${BLOCKLET_COMMAND} config get email`;
     return {
       name: await trimServerOutputVersion(name?.trim()),
       email: await trimServerOutputVersion(email?.trim()),